ELMU #146 Merge setting and settings controller
